chore(coremlit/deps): windit 0.2 -> 0.3, and publish the vector smoother (#98 item 1)
windit 0.3.0 carries three breaking changes coremlit is exposed to plus the
feature #98 item 1 asked for.

The one SOURCE break the compiler found is `EmaRenormalized::new` now taking
`C: Real` (sealed to `f64`): the two sites that passed an `f32` VARIABLE stop
compiling. Every `coverage()` site absorbed the `f32 -> f64` return silently,
because a float literal infers to whatever it is compared against.

`AggregatePolicyKind::EmaRenormalized { alpha }` becomes `f64`, matching what
windit chose for its own wire selector and for the same reason: this enum is the
WIRE type, deserialized from a config surface before any embedding — and so
before a compute scalar — exists, and a decimal in a file has no compute domain
of its own. `aggregate` widens clap's `f32` storage to `f64` before folding a
single component, so an `f32` field resolved the weight on a `2^-24` grid inside
a sum that rounds at `2^-53`. The JSON form is unchanged (a number carries no
width); a configured `0.3` now reaches the fold at full `f64` precision, which
moves an EMA aggregate in its eighth significant digit. `f64::from(0.3f32)`
reproduces the old weights bit for bit.

The `MostCovered` doctest is the downstream-implementor case windit's changelog
warns about — the object-safe `aggregate_values` seam now takes `coverages:
&[f64]` — and is the reason the `--doc` row is the one that proves this break
fixed.

`Real` gained a `Debug` supertrait and a `from_f64` associated function, both
source-breaking for method resolution (E0034) at a call site bounded on `Real`
plus a local trait. coremlit has no `: Real` bound, no `windit::scalar` import
and no `from_f64`/`from_f32` call anywhere, so it is a no-op here.

#98 item 1: `embeddings::clap::smooth` re-exports windit's smoothing tier —
`Smoother` / `SmoothPolicy`, `VectorEma`, `VectorEmaState`, `Identity` — and adds
the clap-typed `smooth()` wrapper, the streaming sibling of `aggregate()`. The
vector smoother, the two traits and the wrapper are flattened onto
`embeddings::clap` the way the aggregation half is; `Identity` and
`VectorEmaState` stay at the module path (a bare `Identity` collides too easily
in a flat namespace). `Ema` / `CadenceEma` are deliberately absent: they are
`Smoother<f32>` scalar low-passes, the wrong shape for a 512-wide embedding, and
re-exporting them would ship a misleading knob — the `SaliencyWeighted`
precedent.

Nine tests drive the smoother over a real `Embedding` stream through coremlit's
own public surface. The load-bearing one is differential: at the same alpha,
output window `i` carries the direction `EmaRenormalized` folds over the prefix
`[0..=i]`, checked at every prefix and three alphas. The rest pin span
preservation, the alpha edges, batch == streaming, fresh-state-per-call, the
empty stream (deliberately not `aggregate`'s `EmptyWindows`), and the determinacy
gate arriving as clap's `Windowing(NonFinite)`.

`window/tests.rs`'s coverage epsilon is tightened to exact equality: at a window
under `2^53`, `coverage()` is the IEEE division of the two exact counts, so the
`1e-7` slack sized for the old `f32` return is now unfalsifiable.

+++ crates/coremlit/src/embeddings/clap/aggregate/mod.rs
@@ -59,9 +59,11 @@ mod tests;
 /// The set is open. windit's trait is slice-level — values arrive already
 /// widened to the `f64` compute domain and unit-normalized, and [`aggregate`]
 /// reconstructs the [`Embedding`] from what the policy returns — so a custom
-/// policy implements [`AggregatePolicy`] over `&[&[f64]]`. Here one that trusts
-/// only the highest-coverage window, exercised through the public seam, no model
-/// required:
+/// policy implements [`AggregatePolicy`] over `&[&[f64]]`, and reads the window
+/// coverages as `&[f64]` — the domain [`Span::coverage`](crate::embeddings::clap::window::Span::coverage)
+/// itself resolves in, so nothing the fold multiplies an embedding by is rounded
+/// through a narrower grid first. Here one that trusts only the highest-coverage
+/// window, exercised through the public seam, no model required:
 ///
 /// ```
 /// use coremlit::embeddings::clap::aggregate::{AggregatePolicy, aggregate};
@@ -75,7 +77,7 @@ mod tests;
 ///     fn aggregate_values(
 ///         &self,
 ///         embeddings: &[&[f64]],
-///         coverages: &[f32],
+///         coverages: &[f64],
 ///         dim: usize,
 ///     ) -> Result<Vec<f64>, WinditError> {
 ///         let (best, _) = coverages
@@ -155,7 +157,20 @@ pub enum AggregatePolicyKind {
   /// Selects [`EmaRenormalized`] with the given smoothing factor.
   EmaRenormalized {
     /// The EMA smoothing factor, forwarded to [`EmaRenormalized::new`].
-    alpha: f32,
+    ///
+    /// `f64`, matching the compute domain the factor multiplies: this enum is
+    /// the *wire* type, deserialized from a config surface before any embedding
+    /// exists, and a decimal in a file has no compute domain of its own.
+    /// [`aggregate`] widens clap's `f32` storage to `f64` before folding a
+    /// single component, so an `f32` field here would have resolved the weight
+    /// at `2^-24` inside a sum that rounds at `2^-53` — the defect windit fixed
+    /// in its own selector. The JSON form is unchanged (a number carries no
+    /// width), so configuration files round-trip exactly as before; a configured
+    /// `0.3` now reaches the fold at full `f64` precision rather than through
+    /// the `f32` grid, which moves an EMA aggregate in its eighth significant
+    /// digit. Pass `f64::from(0.3f32)` to reproduce the pre-0.3 weights bit for
+    /// bit.
+    alpha: f64,
   },
   /// Selects [`CoverageWeightedMean`].
   CoverageWeightedMean,
